test: split skill workshop list coverage
